Make all sends non-blocking
Here we replaced blocking sends in "send_completed" for non-root
processes to non-blocking sends treating the processes in the same
way as the root process.

In "dispatcher" mpi_wait for non-root nodes were replaced by
mpi_test treating them in the same way as the root process.

We are testing if this has any implications on the performace.